Update year length to the actual Gregorian length
The Gregorian Calendar, the one we generally use for day to day, has an average year length of 365.2425 days. This is because the three rules for counting leap years are:

- year % 4 == 0: leap
- year % 100 == 0: normal
- year % 400 == 0: leap

Which is why 2000, for example, was a leap year. So it would be a good idea to update this number to be more accurate.

On a side note, the value before this commit, 365.24, _is_ accurate for the Julian Calendar, on which the Gregorian is based. It just started getting out of sync with the seasons (Easter was creeping closer to June), so a revision was commissioned, and now here we are!